Merging hand tracking L + tap gesture to main (#14)
* environment setup

* HandTracking file addition

* Setup handtracking

Messing around with handtracking

* Update AppModel.swift

* Partially functional tracking of hand joints

Added some joint tracking of the hand. I think all the models for the tracked joints are spawning at the same origin point an not moving though? Not sure yet if it is that, the model just aren't updating, or if the model and values aren't updating.

* Sphere model tracking for hand joints.

Got the joints to be tracked! Purple spheres track your different joints in the immersive space.

* L gesture tracking for either hand

Capabilities for detecting L shaped gesture with either hand for frame spawning. Want to map this to actual Model spawning next.

* L shaped + tap almost

L shaped + tap not working right, but think we are close

* merged fram spawning into test

merged frame spawning into test

* Tap Gesture with L works

Wwooooooooo, need to spawn in objects from it now

* L pose + tap sends notif

Gesture sends single notif to HandTrackingRealityView

* L + tap spawns objects

WOOOOOO

* todos for tmmrw

* Testing new frame spawn

* Fixed appModel reference

Needed to use the @Environment binding and pass in the appModel instead
of using @State

* buggy approach to preframe spawning upon L gesture

When just L gesture is present, a frame object tracks to hand but it is oriented incorrectly and does't go away as intended. Going to pivot on implementation of spawn to a slightly different gesture.

* frame spawning without the purple balls

* L gesture + tap working and looking better

-Got rid of test spheres that tracked to all bone joints
-Added green orb indicators that turn on and off for the L shaped hand indicating gesture
-Reorganized some of the handtracking files

---------

Co-authored-by: Kenneth Yang <kjy5@uw.edu>